add CC1350 SensorTag support
the code was using the deviceName to choose the device implementation based on
the connected hardware's name. So the sensor-tag-2 was changed to specify
CC2650 in its deviceName.

It might be good to also put the name in the filters. This way if a developer chooses
to only use one of these two device implementations only the supported devices
will be listed in the browser device selection dialog