feat: Added "Record Data Trials" experiment [PT-186741111]
This adds the foundation of the future support of logging the data from any GDX device as a time series.

NOTE: this does not yet support connecting to any GDX sensor or logging the data as a time series.  Those are in upcoming stories.

NOTE 2: this also temporarily disables the remote experiment update as that would wipe out this new experiment.