fix: Authored state lookup in authoring mode [PT-187809702]
This fixes a couple of issues found when testing out authoring:

In authoring mode the authoredState comes in as a string which was causing the authoring preview to show an error that the experiment was not selected.

There was still a reference to the "Save Authored State" button, which I believe came from the S3 authoring version.