fix: Stop sensor before restarting recording [PT-187900630]
The GDX sensor was not being stopped before restarting recording which caused erroneous values to be displayed when the sample rate was lowered as they would immediately be recorded.

This also updates the spark graph label to start at time 0.  This was evident at the 0.1 sampler rate as the label was immediately showing 10 seconds at the start.