fix: Stop series data not saving when max values reached [PT-187949831]
Previously there was a useCallback() wrapper on the stop time series function that depended on the formData but this caused issues with the formData being captured as an empty array at the start of multiple runs where the runs hit their max number of values.

This fix adds a ref to the formData so that the closure is not created.